+++ js/changes.md
@@ -6,6 +6,7 @@
 1. Improve `TGraph` update
 1. When draw TH2/TF2 with "surf same" draw option, only lines are drawn
 1. Implement new "POLF" and "POLN" draw options for polar coordinates for `TH2`
+1. Adjust automatic colors handling in `THStack` and `TMultiGraph`
 1. Use "JSROOT" label for main pad button, hide it after 5 seconds from the drawing
 1. Let preserve `BigInt` values when calling `rntupleProcess()` function #418
 1. Use `8.6g`/`10.8g` format for float/double historgam content-to-text conversion in tooltips